backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dateutil==2.9.0
email-validator==2.3.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# OCR
pytesseract==0.3.13
Pillow==11.0.0
opencv-python==4.10.0.84

# Image processing
numpy==2.1.3

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Utilities
python-dotenv==1.0.1

